move UpdateGeometryCache to go.graphics
